Properties JSON First Iteration
Each property currently has 8 values:

id = serves both as a unique identifier, and the order that the property appears on the board.
name = name
lat = station latitude
long = station longitude
value = base value for the property
improvement = the level of improvement on the property, value must be 0,1,2, or 3.
owner = the id of the team that owns the property. If unowned, the owner value is -1.
type = the type of station. 0 is go, 1 is a normal property, 2 is a wildcard (chance, community chest), 3 is a free space (like visiting jail or free parking), and 4 is reserved only for the "go to jail" space.